FoodVisionGuard_starter/FoodVisionGuard/requirements.txt
torch>=2.1
torchvision>=0.16
timm>=0.9.12
albumentations>=1.4.5
opencv-python>=4.9.0.80
numpy>=1.26
pandas>=2.2
matplotlib>=3.8
pytorch-lightning>=2.2.0
scikit-learn>=1.4
pyyaml>=6.0
scikit-image>=0.22
onnx>=1.15
onnxruntime>=1.16
